+++ src/main/java/net/Gabou/projectatmosphere/seasons/SeasonStage.java
@@ -1,0 +1,9 @@
+package net.Gabou.projectatmosphere.seasons;
+
+public enum SeasonStage {
+    SPRING,
+    SUMMER,
+    AUTUMN,
+    WINTER,
+    NEUTRAL
+}